requirement.txt
pip==21.1.2
numpy==1.19.5
pandas==1.1.5
scikit-learn==0.24.2
PyYAML==6.0